# target/classes/swagger/adexchange.yaml
...
swagger: '2.0'
info:
  title: TechM Arena Ad Exchange API
  description: API specification for TechM Arena Ad Exchange
  version: "1.0.0"
host: api.test.com
schemes:
  - http
  - https
basePath: /v1
produces:
  - application/json
paths:
  /user:
    get:
      description: |
        List of User
      parameters:
        - name: pageNo
          in: query
          description: pageNo
          required: false
          type: integer
        - name: pageSize
          in: query
          description: pageSize
          required: false
          type: integer
      tags:
        - User
      responses:
        200:
          description: List of User
          schema:
            $ref: '#/definitions/UserList'          
    post:
      description: Creates a new User.
      operationId: createUser
      parameters:
        - name: user
          in: body
          description: User to be created
          required: true
          schema:
            $ref: '#/definitions/User'
      responses:
        '200':
          description: created User
          schema:
            $ref: '#/definitions/User'
    put:
      description: Updates the User.
      operationId: updateUser
      parameters:
        - name: user
          in: body
          description: User to be updated
          required: true
          schema:
            $ref: '#/definitions/User'
      responses:
        '200':
          description: updated User
          schema:
            $ref: '#/definitions/User'
  /user/{id}:
    get:
      description: Returns User for the given id
      operationId: getUser
      parameters:
        - name: id
          in: path
          description: ID of User
          required: true
          type: integer
      responses:
        '200':
          description: User response
          schema:
            $ref: '#/definitions/User'
    delete:
      description: deletes a single User based on the ID supplied
      operationId: deleteUser
      parameters:
        - name: id
          in: path
          description: ID of User to delete
          required: true
          type: integer
      responses:
        '200':
          description: User deleted
  /user/search:
    get:
      description: Search User
      operationId: searchUser
      parameters:
        - name: email
          in: query
          description: email
          required: false
          type: string
        - name: loginName
          in: query
          description: loginName
          required: false
          type: string
        - name: password
          in: query
          description: password
          required: false
          type: string
        - name: status
          in: query
          description: status
          required: false
          type: integer
        - name: userType
          in: query
          description: userType
          required: false
          type: integer
        - name: hintQuestion1
          in: query
          description: hintQuestion1
          required: false
          type: string
        - name: hintAnswer1
          in: query
          description: hintAnswer1
          required: false
          type: string
        - name: hintQuestion2
          in: query
          description: hintQuestion2
          required: false
          type: string
        - name: hintAnswer2
          in: query
          description: hintAnswer2
          required: false
          type: string
        - name: isAdmin
          in: query
          description: isAdmin
          required: false
          type: boolean
        - name: createdTimeStart
          in: query
          description: createdTimeStart
          required: false
          type: string
        - name: createdTimeEnd
          in: query
          description: createdTimeEnd
          required: false
          type: string
        - name: updatedTimeStart
          in: query
          description: updatedTimeStart
          required: false
          type: string
        - name: updatedTimeEnd
          in: query
          description: updatedTimeEnd
          required: false
          type: string
        - name: createdByUser
          in: query
          description: createdByUser
          required: false
          type: string
        - name: updatedByUser
          in: query
          description: updatedByUser
          required: false
          type: string
        - name: advertiserId
          in: query
          description: advertiserId
          required: false
          type: String
        - name: publisherId
          in: query
          description: publisherId
          required: false
          type: String
        - name: roleId
          in: query
          description: roleId
          required: false
          type: String
        - name: pageNo
          in: query
          description: pageNo
          required: false
          type: integer
        - name: pageSize
          in: query
          description: pageSize
          required: false
          type: integer
      responses:
        '200':
          description: List of User
          schema:
            $ref: '#/definitions/UserList'
  /user/search/keyword:
    get:
      description: Search User by keyword
      operationId: searchUserByKeyword
      parameters:
        - name: pageNo
          in: query
          description: pageNo
          required: false
          type: integer
        - name: pageSize
          in: query
          description: pageSize
          required: false
          type: integer
      responses:
        '200':
          description: List of User
          schema:
            $ref: '#/definitions/UserList'            
...
  /video:
    get:
      description: |
        List of Video
      parameters:
        - name: pageNo
          in: query
          description: pageNo
          required: false
          type: integer
        - name: pageSize
          in: query
          description: pageSize
          required: false
          type: integer
      tags:
        - Campaign
      responses:
        200:
          description: List of Video
          schema:
            $ref: '#/definitions/VideoList'          
    post:
      description: Creates a new Video.
      operationId: createVideo
      parameters:
        - name: video
          in: body
          description: Video to be created
          required: true
          schema:
            $ref: '#/definitions/Video'
      responses:
        '200':
          description: created Video
          schema:
            $ref: '#/definitions/Video'
    put:
      description: Updates the Video.
      operationId: updateVideo
      parameters:
        - name: video
          in: body
          description: Video to be updated
          required: true
          schema:
            $ref: '#/definitions/Video'
      responses:
        '200':
          description: updated Video
          schema:
            $ref: '#/definitions/Video'
...
definitions:
  User:
    type: object
    properties:
      id:
        type: integer
        description: Id
      email:
        type: string
      loginName:
        type: string
      password:
        type: string
      status:
        type: integer
      userType:
        type: integer
      hintQuestion1:
        type: string
      hintAnswer1:
        type: string
      hintQuestion2:
        type: string
      hintAnswer2:
        type: string
      isAdmin:
        type: boolean
      createdTime:
        type: string
        format: date
      updatedTime:
        type: string
        format: date
      createdByUser:
        type: string
      updatedByUser:
        type: string
      advertiser:
        type: Advertiser
      publisher:
        type: Publisher
      role:
        type: Role
  UserList:
    type: object
    properties:
      count:
        type: integer
        format: int64
        description: Total count of objects
      users:
        type: array
        items:
          $ref: '#/definitions/User'      
...
  Video:
    type: object
    properties:
      id:
        type: integer
        description: Id
      title:
        type: string
      videoUrl:
        type: string
      videoType:
        type: string
      delivery:
        type: string
      bitrate:
        type: string
      width:
        type: integer
      height:
        type: integer
      duration:
        type: number
        format: float
  VideoList:
    type: object
    properties:
      count:
        type: integer
        format: int64
        description: Total count of objects
      videos:
        type: array
        items:
          $ref: '#/definitions/Video'      
...